+++ src/interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from './dateTime/datetime.interface';
+export * from './dateTimeInput/dateTimeInput.interface';