+++ src/legacy/picker/components/dayPicker/dayPicker.component.ts
@@ -1,7 +1,7 @@
 import {Component, ChangeDetectionStrategy, Inject, ChangeDetectorRef} from '@angular/core';
 import {extend} from '@jscrpt/common';
 
-import {DateTimeValue} from '../../../../misc/datetime.interface';
+import {DateTimeValue} from '../../../../interfaces/dateTime/datetime.interface';
 import {DATE_API} from '../../../../misc/tokens';
 import {DateApi, DateApiObject} from '../../../../services/dateApi.interface';
 import {DateTimePicker, DayData, PeriodData} from '../../misc/datetimePicker.interface';