+++ src/legacy/picker/components/pickerBase.component.ts
@@ -1,6 +1,6 @@
 import {ChangeDetectorRef, Directive, Inject} from '@angular/core';
 
-import {DateTimeValue} from '../../../misc/datetime.interface';
+import {DateTimeValue} from '../../../interfaces/dateTime/datetime.interface';
 import {DATE_API} from '../../../misc/tokens';
 import {DateApi, DateApiObject} from '../../../services/dateApi.interface';
 import {DateTimePicker, PeriodData} from '../misc/datetimePicker.interface';