+++ src/legacy/picker/components/pickerImplBase.component.ts
@@ -2,7 +2,7 @@ import {ChangeDetectorRef, Directive, HostListener, Inject} from '@angular/core'
 import {extend} from '@jscrpt/common';
 import {Observable, Subject} from 'rxjs';
 
-import {DateTimeValue} from '../../../misc/datetime.interface';
+import {DateTimeValue} from '../../../interfaces/dateTime/datetime.interface';
 import {DATE_API} from '../../../misc/tokens';
 import {DateApi, DateApiObject} from '../../../services/dateApi.interface';
 import {DateTimePicker} from '../misc/datetimePicker.interface';