+++ src/modules/dateTime/directives/index.ts
@@ -1,0 +1,2 @@
+export * from './dateTimeControlValueAccessor/dateTimeControlValueAccessor.directive';
+export * from './dateTimeInput/dateTimeInput.directive';