+++ src/modules/index.ts
@@ -1,0 +1,2 @@
+export * from './datePipes.module';
+export * from './dateTime';